requirements.txt
flask==2.3.3
numpy==1.24.3
requests==2.31.0
click==8.1.7
python-dotenv==1.0.0
pydantic==2.5.0
